refactor: moved subCateg types to external file
