+++ src/types/types.ts
@@ -4,11 +4,19 @@ export interface Category {
   flag: string
 }
 
-export interface CategoryState {
+
+type CommonCategorySet = {
   categories: Category[],
-  category: Category | undefined,
+  category: Category | undefined | Category[] | [],
   highlight: number,
   open: boolean,
+}
+export interface CategoryState extends CommonCategorySet{
+  category: Category | undefined,
+}
+
+export interface SubCategoryState extends CommonCategorySet {
+  category: Category[] | []
 }
 
 export enum CategoryActionTypes {
@@ -46,3 +54,48 @@ export type CategoryAction =
   DeleteCategoryAction    |
   HighlightCategoryAction |
   ToggleCategoryAction
+
+// ==========
+
+
+export enum SubCategoryActionTypes {
+  GET_SUB_CATEGORIES      = 'GET_SUB_CATEGORIES',
+  SET_SUB_CATEGORIES      = 'SET_SUB_CATEGORIES',
+  DELETE_SUB_CATEGORY     = 'DELETE_SUB_CATEGORY',
+  HIGHLIGHT_SUB_CATEGORY  = 'HIGHLIGHT_SUB_CATEGORY',
+  TOGGLE_SUB_CATEGORY     = 'TOGGLE_SUB_CATEGORY',
+  DELETE_ONE_SUB_CATEGORY = 'DELETE_ONE_SUB_CATEGORY',
+}
+
+interface GetSubCategoriesAction {
+  type: SubCategoryActionTypes.GET_SUB_CATEGORIES;
+  payload: Category[]
+}
+interface SetSubCategoriesAction {
+  type: SubCategoryActionTypes.SET_SUB_CATEGORIES;
+  payload: Category
+}
+interface DeleteSubCategoryAction {
+  type: SubCategoryActionTypes.DELETE_SUB_CATEGORY;
+  payload: Category
+}
+interface HighlightSubCategoryAction {
+  type: SubCategoryActionTypes.HIGHLIGHT_SUB_CATEGORY;
+  payload: number
+}
+interface ToggleSubCategoryAction {
+  type: SubCategoryActionTypes.TOGGLE_SUB_CATEGORY;
+  payload: Category
+}
+interface DeleteOneSubCategoryAction {
+  type: SubCategoryActionTypes.DELETE_ONE_SUB_CATEGORY;
+  payload: Category
+}
+
+export type SubCategoryAction =
+  GetSubCategoriesAction     |
+  SetSubCategoriesAction     |
+  DeleteSubCategoryAction    |
+  HighlightSubCategoryAction |
+  ToggleSubCategoryAction    |
+  DeleteOneSubCategoryAction
